[raw.orca] move SIS3316 code to orca_digitizers.py and tidy up
* rename ORSIS3302DecoderForEnergy to ORSIS3302WaveformDecoder
* add missing type annotations and fix docstrings
* remove commented code at the end of orca_streamer.py
