+++ src/pygama/raw/orca/orca_streamer.py
@@ -14,7 +14,10 @@ from pygama.raw.orca.orca_flashcam import (  # noqa: F401
     ORFlashCamListenerConfigDecoder,
 )
 from pygama.raw.orca.orca_header_decoder import OrcaHeaderDecoder
-from pygama.raw.orca.orca_SIS3316 import ORSIS3316WaveformDecoder  # noqa: F401
+from pygama.raw.orca.orca_digitizers import (  # noqa: F401
+    ORSIS3302WaveformDecoder,
+    ORSIS3316WaveformDecoder,
+)
 from pygama.raw.raw_buffer import RawBuffer, RawBufferLibrary
 
 log = logging.getLogger(__name__)
@@ -284,197 +287,3 @@ class OrcaStreamer(DataStreamer):
         rbl = self.rbl_id_dict[data_id]
         self.any_full |= decoder.decode_packet(packet, self.packet_id, rbl)
         return True
-
-        '''
-# Import orca_digitizers so that the list of OrcaDecoder.__subclasses__ gets populated
-# Do it here so that orca_digitizers can import the functions above here
-from . import orca_digitizers, orca_flashcam
-
-def process_orca(daq_filename, raw_file_pattern, n_max=np.inf, ch_groups_dict=None, verbose=False, buffer_size=1024):
-    """
-    convert ORCA DAQ data to "raw" lh5
-
-    ch_groups_dict: keyed by decoder_name
-    """
-    lh5_store = lh5.Store()
-
-    f_in = open_orca(daq_filename)
-    if f_in == None:
-        print("Couldn't find the file %s" % daq_filename)
-        sys.exit(0)
-
-    # parse the header. save the length so we can jump past it later
-    reclen, header_nbytes, header_dict = parse_header(daq_filename)
-
-    # figure out the total size
-    SEEK_END = 2
-    f_in.seek(0, SEEK_END)
-    file_size = float(f_in.tell())
-    f_in.seek(0, 0)  # rewind
-    file_size_MB = file_size / 1e6
-    print("Total file size: {:.3f} MB".format(file_size_MB))
-    print("Run number:", get_run_number(header_dict))
-
-
-    # Build the dict used in the inner loop for passing data packets to decoders
-    decoders = {}
-
-    # First build a list of all decoder names that might be in the data
-    # This is a dict of names keyed off of data_id
-    id2dn_dict = get_id_to_decoder_name_dict(header_dict)
-    if verbose:
-        print("Data IDs present in ORCA file header are:")
-        for data_id in id2dn_dict:
-            print(f"    {data_id}: {id2dn_dict[data_id]}")
-
-    # Invert the previous list, to get a list of decoder ids keyed off of
-    # decoder names
-    dn2id_dict = {name:data_id for data_id, name in id2dn_dict.items()}
-
-    # By default we decode all data for which we have decoders. If the user
-    # provides a ch_group_dict, we will only decode data from decoders keyed in
-    # the dict.
-    decode_all_data = True
-    decoders_to_run = dn2id_dict.keys()
-    if ch_groups_dict is not None:
-        decode_all_data = False
-        decoders_to_run = ch_groups_dict.keys()
-
-    # Now get the actual requested decoders
-    for sub in OrcaDecoder.__subclasses__():
-        decoder = sub() # instantiate the class
-        if decoder.decoder_name in decoders_to_run:
-            decoder.dataID = dn2id_dict[decoder.decoder_name]
-            decoder.set_header_dict(header_dict)
-            decoders[decoder.dataID] = decoder
-    if len(decoders) == 0:
-        print("No decoders. Exiting...")
-        sys.exit(1)
-    if verbose:
-        print("pygama will run these decoders:")
-        for data_id, dec in decoders.items():
-            print("   ", dec.decoder_name+ ", id =", data_id)
-
-    # Now cull the decoders_to_run list
-    new_dtr = []
-    for decoder_name in decoders_to_run:
-        data_id = dn2id_dict[decoder_name]
-        if data_id not in decoders.keys():
-            print("warning: no decoder exists for", decoder_name, "... will skip its data.")
-        else: new_dtr.append(decoder_name)
-    decoders_to_run = new_dtr
-
-    # prepare ch groups
-    if ch_groups_dict is None:
-        ch_groups_dict = {}
-        for decoder_name in decoders_to_run:
-            ch_groups = create_dummy_ch_groups()
-            ch_groups_dict[decoder_name] = ch_groups
-            grp_path_template = f'{decoder_name}/raw'
-            set_outputs(ch_groups, out_file_template=raw_file_pattern, grp_path_template=grp_path_template)
-    else:
-        for decoder_name, ch_groups in ch_groups_dict.items():
-            expand_ch_groups(ch_groups)
-            set_outputs(ch_groups, out_file_template=raw_file_pattern, grp_path_template='{system}/{group_name}/raw')
-
-    # Set up tables for data
-    ch_tables_dict = {}
-    for data_id, dec in decoders.items():
-        decoder_name = id2dn_dict[data_id]
-        ch_groups = ch_groups_dict[decoder_name]
-        ch_tables_dict[data_id] = build_tables(ch_groups, buffer_size, dec)
-    max_tbl_size = 0
-
-    # -- scan over raw data --
-    print("Beginning daq-to-raw processing ...")
-
-    packet_id = 0  # number of events decoded
-    unrecognized_data_ids = []
-
-    # skip the header using reclen from before
-    # reclen is in number of longs, and we want to skip a number of bytes
-    f_in.seek(reclen * 4)
-
-    n_entries = 0
-    unit = "B"
-    if n_max < np.inf and n_max > 0:
-        n_entries = n_max
-        unit = "id"
-    else:
-        n_entries = file_size
-    progress_bar = tqdm_range(0, int(n_entries), text="Processing", verbose=verbose, unit=unit)
-    file_position = 0
-
-    # start scanning
-    while (packet_id < n_max and f_in.tell() < file_size):
-        packet_id += 1
-
-        try:
-            packet, data_id = get_next_packet(f_in)
-        except EOFError:
-            break
-        except Exception as e:
-            print("Failed to get the next event ... Exception:", e)
-            break
-
-        if decode_all_data and data_id not in decoders:
-            if data_id not in unrecognized_data_ids:
-                unrecognized_data_ids.append(data_id)
-            continue
-
-        if data_id not in decoders: continue
-        decoder = decoders[data_id]
-
-        # Clear the tables if the next read could overflow them.
-        # Only have to check this when the max table size is within
-        # max_n_rows_per_packet of being full.
-        if max_tbl_size + decoder.max_n_rows_per_packet() >= buffer_size:
-            ch_groups = ch_groups_dict[id2dn_dict[data_id]]
-            max_tbl_size = 0
-            for group_info in ch_groups.values():
-                tbl = group_info['table']
-                if tbl.is_full():
-                    group_path = group_info['group_path']
-                    out_file = group_info['out_file']
-                    lh5_store.write_object(tbl, group_path, out_file, n_rows=tbl.loc)
-                    tbl.clear()
-                if tbl.loc > max_tbl_size: max_tbl_size = tbl.loc
-        else: max_tbl_size += decoder.max_n_rows_per_packet()
-
-        tables = ch_tables_dict[data_id]
-        decoder.decode_packet(packet, tables, packet_id, header_dict)
-
-        if verbose:
-            if n_max < np.inf and n_max > 0:
-                update_len = 1
-            else:
-                update_len = f_in.tell() - file_position
-                file_position = f_in.tell()
-            update_progress(progress_bar, update_len)
-
-
-    print("Done. Last packet ID:", packet_id)
-    f_in.close()
-
-    # final write to file
-    for dec_name, ch_groups in ch_groups_dict.items():
-        for group_info in ch_groups.values():
-            tbl = group_info['table']
-            if tbl.loc == 0: continue
-            group_path = group_info['group_path']
-            out_file = group_info['out_file']
-            lh5_store.write_object(tbl, group_path, out_file, n_rows=tbl.loc)
-            print('last write')
-            tbl.clear()
-
-    if len(unrecognized_data_ids) > 0:
-        print("WARNING, Found the following unknown data IDs:")
-        for data_id in unrecognized_data_ids:
-            try:
-                print("  {}: {}".format(data_id, id2dn_dict[data_id]))
-            except KeyError:
-                print("  {}: Unknown".format(data_id))
-        print("hopefully they weren't important!\n")
-
-    print("Wrote RAW File:\n    {}\nFILE INFO:".format(raw_file_pattern))
-        '''
